requirements.txt
pynput
pillow
jaraco.clipboard